updated permisions.py and located in shared giftcompare folder, added permission_classes to user views
