--- giftcompare/categories/permissions.py
@@ -1,14 +1,0 @@
-from rest_framework import permissions
-
-class IsAdminOrReadOnly(permissions.BasePermission):
-    """
-    Custom permission to only allow admin users to edit it.
-    """
-    def has_permission(self, request, view):
-        if request.method in permissions.SAFE_METHODS:
-            # The method is a safe method
-            return True
-        else:
-            # The method isn't a safe method
-            # Only allow it if the user is an admin
-            return request.user.is_staff
