+++ giftcompare/users/views.py
@@ -6,15 +6,20 @@ from rest_framework.response import Response
 from rest_framework import status
 from .models import CustomUser
 from .serializers import CustomUserSerializer
-
+from giftcompare.permissions import IsAdminOrReadOnly, IsSelfOrReadOnly
 
 class CustomUserList(APIView):
+    permission_classes = [IsAdminOrReadOnly]
+
     def get(self, request):
         users = CustomUser.objects.all()
         serializer = CustomUserSerializer(users, many=True)
         return Response(serializer.data)
     
     def post(self, request):
+        if request.user.is_authenticated:
+            return Response(status=status.HTTP_403_FORBIDDEN 
+                            , data={"detail": "Not allowed to perform this action"})
         serializer = CustomUserSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -22,6 +27,7 @@ class CustomUserList(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 class CustomUserDetail(APIView):
+    permission_classes = [IsAdminOrReadOnly, IsSelfOrReadOnly]
     def get_object(self, pk):
         try:
             return CustomUser.objects.get(pk=pk)
